# expression: implement vectorized evaluation for `builtinFormatSig`