--- a/expression/builtin_string_vec.go
+++ b/expression/builtin_string_vec.go
@@ -1858,30 +1858,29 @@ func (b *builtinFormatSig) vectorized() bool {
 	return true
 }
 
-// evalString evals FORMAT(X,D).
-// See https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_format
-func (b *builtinFormatSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
+// vecEvalNumDecArgsForFormat evaluates first 2 arguments, i.e, x and d, for function `format`.
+func vecEvalNumDecArgsForFormat(b *baseBuiltinFunc, input *chunk.Chunk) ([]string, []string, []bool, error) {
 	n := input.NumRows()
 	isDecimal := b.args[0].GetType().EvalType() == types.ETDecimal
 	var buf0 *chunk.Column
 	buf0, err := b.bufAllocator.get(b.args[0].GetType().EvalType(), n)
 	if err != nil {
-		return err
+		return nil, nil, nil, err
 	}
 	defer b.bufAllocator.put(buf0)
 	buf1, err := b.bufAllocator.get(types.ETInt, n)
 	if err != nil {
-		return err
+		return nil, nil, nil, err
 	}
 	defer b.bufAllocator.put(buf1)
 	if err := b.args[1].VecEvalInt(b.ctx, input, buf1); err != nil {
-		return err
+		return nil, nil, nil, err
 	}
 
 	xStrBuf := make([]string, n)
 	if isDecimal {
 		if err := b.args[0].VecEvalDecimal(b.ctx, input, buf0); err != nil {
-			return err
+			return nil, nil, nil, err
 		}
 		buf1.MergeNulls(buf0)
 		d64s := buf0.Decimals()
@@ -1893,7 +1892,7 @@ func (b *builtinFormatSig) vecEvalString(input *chunk.Chunk, result *chunk.Colum
 		}
 	} else {
 		if err := b.args[0].VecEvalReal(b.ctx, input, buf0); err != nil {
-			return err
+			return nil, nil, nil, err
 		}
 		buf1.MergeNulls(buf0)
 		f64s := buf0.Float64s()
@@ -1906,14 +1905,35 @@ func (b *builtinFormatSig) vecEvalString(input *chunk.Chunk, result *chunk.Colum
 	}
 
 	d64s := buf1.Int64s()
+	isNulls := make([]bool, n)
+	dStrBuf := make([]string, n)
+	for i := 0; i < n; i++ {
+		if buf1.IsNull(i) {
+			isNulls[i] = true
+			continue
+		}
+		isNulls[i] = false
+		dStrBuf[i] = strconv.FormatInt(d64s[i], 10)
+	}
+	return xStrBuf, dStrBuf, isNulls, nil
+}
+
+// evalString evals FORMAT(X,D).
+// See https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_format
+func (b *builtinFormatSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
+	xStrBuf, dStrBuf, isNulls, err := vecEvalNumDecArgsForFormat(&b.baseBuiltinFunc, input)
+	if err != nil {
+		return err
+	}
+	n := input.NumRows()
 	result.ReserveString(n)
 	var formatString string
 	for i := 0; i < n; i++ {
-		if buf1.IsNull(i) {
+		if isNulls[i] {
 			result.AppendNull()
 			continue
 		}
-		formatString, err = mysql.GetLocaleFormatFunction("en_US")(xStrBuf[i], strconv.FormatInt(d64s[i], 10))
+		formatString, err = mysql.GetLocaleFormatFunction("en_US")(xStrBuf[i], dStrBuf[i])
 		if err != nil {
 			return err
 		}