PCP #12106
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for `builtinFormatSig` , for #12106

### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc-3              2834            441774 ns/op           96545 B/op       5120 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc-3           2229            545643 ns/op           96544 B/op       5120 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc#01-3           2737            449215 ns/op          124608 B/op       5120 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc#01-3        2139            553661 ns/op          124609 B/op       5120 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      5.596s

```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test